fix: handle arguments and defaults consistently
Default function arguments were not being handled the same way as
function arguments. It was possible that an expression that could be
passed as an argument could not have been used as a default value for
the same argument. This patch fixes this issue, and simplifies some of
the parsing code for arguments and function signatures.