feat: simple pretty-printer for KLR
This change adds two related things: a pretty-printer for KLR terms,
and tensor names. Tensor names make the pretty printing nicer, but
have a second purpose. By naming all of the tensors, we can scan a KLR
kernel to collect up all of the input, output, and intermediate
tensors that will be needed to run the kernel. For argument tensors,
the generated tensor names are changed to the argument variable names;
this is just for readability.