feat: extend builtin object capabilities
The main change in this patch is to extend the Object structure with
new functions for indexing and binary operators. In addition, the
operator names have been converted from strings to enumerations, and
the translation of index expressions has been reworked to allow more
flexibility. For instance, the new implementation can now support
`t[f()]` where f is an arbitrary user function or builtin.