CI: Make a wheel
This took a lot of time to figure out, but we can generate
linux and osx wheels now on GitHub Actions that seem functional
on both systems. Will probably hit problems, e.g. with different
Mac OSs, but I suspect that the right tweaks cibuildwheel arguments will
be able to fix those.