feat: Detect undefined symbols
Our current Python parser does not notice undefined symbols.
While we don't do anything with them now, we want to pass
that info back to the caller in the forthcoming linter.
In addition, we intended parsing to fail when there
are undefined symbols, as indicated in test_basic.py, but
it was failing for an unrelated reason before, and broke
a few commits ago.

This change just makes a best effort attempt to notice
undefined symbols, and crashes accordingly in the klr binary.
In the future we will do more with this.

This change fixes the broken CI build.